+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 This is a high-level changelog for each released versions of the plugin.
 For a more detailed list of past and incoming changes, see the commit history.
 
+Master
+------
+- Exclude properties from being printed by using `print_exclude` option
+  
 1.0.1
 ------
 - Sort properties by name alphabetically